gmo: handle chainup correctly
Unlike classic GObject we can't store the parent class
in the static variable, because now the code is agnostic to
it's parent.
So we have to dynamically get the parent class each type.

What we have is:
-----------------
We have an instance of some class.
The code of our mutogene is somewhere in the middle.
So we need to peek the class of the parent of that mutogene exactly
in that case.
To do this we take an instance and iterate through it's class tree
until we find the class that corresponds to our mutogene.
Then we take parent type, and peek it's class. Awesome.